Add winner text below game, disable clicks after game is  over
Separate calcWinner into another file

Update README.md
