+++ README.md
@@ -1,7 +1,6 @@
 # Connect Four
 
 ## To be added:
-* Win Calculation function
 * Responsive website design
 * Game status text
 * Animations
